==> notebooks/T3/regressao/04_Feature_Engineering.ipynb <==
# Engenharia de Features

Este notebook tem como objetivo transformar os dados consolidados em variáveis úteis para o modelo de regressão.

**Principais Transformações:**
* **Cálculo de Cubagem:** Determinação do peso cúbico e do peso final.
* **Cálculo de Distância:** Aplicação da fórmula de Haversine para obter a distância entre vendedor e cliente.
* **Seleção Final:** Filtragem das colunas, mantendo apenas as *features* essenciais e a variável *target*.

import pandas as pd
import numpy as np

from pathlib import Path
import sys
import os

project_root = os.path.abspath("../../..")
if project_root not in sys.path:
    sys.path.append(project_root)

from src.T3.regressao.features import calcular_peso_cubado, haversine_vectorized

BASE_PATH = Path("../../../")
DATA_INTERIM = BASE_PATH / "data" / "interim" / "T3" / "regressao"
DATA_PROCESSED = BASE_PATH / "data" / "processed" / "T3" / "regressao"

dataset_merged = pd.read_csv(DATA_INTERIM / "dataset_merged.csv")
dataset_merged.head()

## Volume e Densidade

df = dataset_merged.copy()
df['volume_cm3'] = df['product_length_cm'] * df['product_height_cm'] * df['product_width_cm']
df['product_density'] = df['product_weight_g'] / (df['volume_cm3'])

## Peso Final

No transporte de cargas, o frete é cobrado com base no maior valor entre o **peso físico** real e o **peso cubado**.

**Fórmula utilizada:**
$$Peso\ Cubado\ (kg) = \frac{Comprimento \times Largura \times Altura}{Fator\ de\ Cubagem}$$

**Critérios adotados:**
* **Fator de Cubagem (6000):** É o padrão oficial da IATA e utilizado pela maioria das transportadoras rodoviárias e Correios no Brasil para converter volume ($cm^3$) em peso taxável ($kg$).
* **Conversão de Unidade:** A fórmula original resulta em quilogramas. Como a coluna de peso original (`product_weight_g`) está em gramas, multiplicamos o resultado do peso cúbico por **1000** para manter a consistência das unidades.
* **Peso Final:** Definido como o valor máximo entre o peso real e o peso cúbico calculado (`max(peso_real, peso_cubado)`).

df['peso_cubado_g'] = calcular_peso_cubado(
    df['product_length_cm'], 
    df['product_height_cm'], 
    df['product_width_cm'], 
    fator_cubagem=6000
)

df['final_weight_g'] = np.maximum(df['product_weight_g'], df['peso_cubado_g'])

## Comprador e vendedor são do mesmo estado?

df['same_state'] = (df['customer_state'] == df['seller_state']).astype(int)

## Distância

Para estimar a distância de entrega, utilizamos a **Fórmula de Haversine**.

**Fórmula utilizada:**
$$d = 2r \arcsin\left(\sqrt{\sin^2\left(\frac{\Delta\phi}{2}\right) + \cos(\phi_1)\cos(\phi_2)\sin^2\left(\frac{\Delta\lambda}{2}\right)}\right)$$

Onde:
* $r$: Raio da Terra (~6371 km)
* $\phi$: Latitude (em radianos)
* $\lambda$: Longitude (em radianos)

df['distance_km'] = haversine_vectorized(
    df['lat_seller'].values, 
    df['lng_seller'].values, 
    df['lat_customer'].values, 
    df['lng_customer'].values,
    R=6371.0 # Raio da terra em KM
)

## Seleção de Features e Target

cols_finais = [
    'price',
    'final_weight_g',
    'volume_cm3',
    'product_density',
    'customer_state',
    'seller_state',
    'same_state',
    'lat_customer',
    'lng_customer',
    'lat_seller',
    'lng_seller',
    'distance_km',
    'freight_value'
]

df_final = df[cols_finais].copy()
df_final.head()

## Exportação dos Dados

df_final.to_csv(DATA_PROCESSED / "freight_prediction_abt.csv", index=False)
